add security layer

+++ server/wutsi-blog-service/src/main/kotlin/com/wutsi/blog/story/endpoint/DeleteStoryCommandExecutor.kt
@@ -1,5 +1,6 @@
 package com.wutsi.blog.story.endpoint
 
+import com.wutsi.blog.security.service.SecurityManager
 import com.wutsi.blog.story.dto.DeleteStoryCommand
 import com.wutsi.blog.story.service.StoryService
 import org.springframework.web.bind.annotation.PostMapping
@@ -12,8 +13,12 @@ import javax.validation.Valid
 @RequestMapping("/v1/stories/commands/delete")
 class DeleteStoryCommandExecutor(
     private val storyService: StoryService,
+    private val securityManager: SecurityManager,
 ) {
     @PostMapping()
-    fun create(@RequestBody @Valid command: DeleteStoryCommand) =
+    fun create(@RequestBody @Valid command: DeleteStoryCommand) {
+        securityManager.checkStoryOwnership(command.storyId)
+
         storyService.delete(command)
+    }
 }
